composer scripts内でphpを使わないように変更
