+++ composer.json
@@ -30,31 +30,6 @@
     "scripts":{
          "darwin-chmod": ["chmod -R 777 src/data", "chmod 777 src/Asazuke.log desktop.sql"],
          "win32-copy-conf": ["copyConf.bat"],
-         "which-php": ["php -r 'echo $_SERVER[\"_\"];'"],
-         "run"          : ["php index.php"],
-         "run:file-sql"      : ["php index.php file-sql"],
-         "run:file-sql-json" : ["php index.php file-sql-json"],
-         "run:conf-json": ["php index.php conf-json"],
-         "run:site-scan0": ["php index.php site-scan0"],
-         "run:site-scan" : ["php index.php site-scan"],
-         "run:csv"      : ["php index.php csv"],
-         "run:show"     : ["php index.php show"],
-         "run:table"    : ["php index.php table"],
-         "run:schema"   : ["php index.php schema"],
-         "run:rm-dat"   : ["php index.php rm-dat"],
-         "run:rm-html"  : ["php index.php rm-html"],
-         "run:site-validation"      : ["php index.php rm-dat", "php index.php rm-html", "php index.php site-validation"],
-         "run:site-validation-ex"   : ["php index.php rm-dat", "php index.php rm-html", "php index.php site-validation", "php index.php rm-cssworks", "php index.php cssworks"],
-         "run:site-validation-show" : ["php index.php site-validation-show"],
-         "run:site-validation-json" : ["php index.php site-validation-json"],
-         "run:site-validation-csv"  : [ "php index.php site-validation-csv"],
-         "run:site-validation-csv-origin"  : ["php index.php site-validation-csv-origin"],
-         "run:site-validation-result"  : ["php index.php site-validation-result"],
-         "run:rm-scraping": ["php index.php rm-scraping"],
-         "run:scraping"   : ["php index.php rm-scraping", "php index.php scraping"],
-         "run:rm-cssworks": ["php index.php rm-cssworks"],
-         "run:cssworks"  :  ["php index.php rm-cssworks", "php index.php cssworks"],
-         "run:inline":  ["php index.php inline"],
          "test" : ["vendor/phpunit/phpunit/phpunit --configuration phpunit.xml"]
     },
     "autoload" : {
